Moved to CppUTest for IRQ testing
No test has been changed, but unit tests are now grouped using CppUTest macros
