+++ module.json
@@ -30,5 +30,8 @@
   "dependencies": {
     "mbed": "*"
   },
+  "testDependencies": {
+    "mbed-cpputest": "*"
+  },
   "targetDependencies": {}
 }
