+++ test/irq/common.h
@@ -1,0 +1,18 @@
+/***************************************************************
+ * This confidential and  proprietary  software may be used only
+ * as authorised  by  a licensing  agreement  from  ARM  Limited
+ *
+ *             (C) COPYRIGHT 2013-2015 ARM Limited
+ *                      ALL RIGHTS RESERVED
+ *
+ *  The entire notice above must be reproduced on all authorised
+ *  copies and copies  may only be made to the  extent permitted
+ *  by a licensing agreement from ARM Limited.
+ *
+ ***************************************************************/
+#ifndef __COMMON_H__
+#define __COMMON_H__
+
+extern volatile int g_flag;
+
+#endif/*__COMMON_H__*/
